LPS-200037 move methods from AssetRSSHelper to the only usage in AssetPublisherPortlet, rename methods with the same name, and removing not used methods
