+++ modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/portlet/AssetPublisherPortlet.java
@@ -7,6 +7,7 @@ package com.liferay.asset.publisher.web.internal.portlet;
 
 import com.liferay.asset.constants.AssetWebKeys;
 import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
+import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.model.ClassType;
 import com.liferay.asset.kernel.model.ClassTypeField;
@@ -15,11 +16,13 @@ import com.liferay.asset.list.asset.entry.provider.AssetListAssetEntryProvider;
 import com.liferay.asset.list.service.AssetListEntrySegmentsEntryRelLocalService;
 import com.liferay.asset.publisher.constants.AssetPublisherPortletKeys;
 import com.liferay.asset.publisher.constants.AssetPublisherWebKeys;
+import com.liferay.asset.publisher.util.AssetEntryResult;
 import com.liferay.asset.publisher.util.AssetPublisherHelper;
+import com.liferay.asset.publisher.web.internal.configuration.AssetPublisherSelectionStyleConfigurationUtil;
 import com.liferay.asset.publisher.web.internal.configuration.AssetPublisherWebConfiguration;
+import com.liferay.asset.publisher.web.internal.constants.AssetPublisherSelectionStyleConstants;
 import com.liferay.asset.publisher.web.internal.display.context.AssetPublisherDisplayContext;
 import com.liferay.asset.publisher.web.internal.helper.AssetPublisherWebHelper;
-import com.liferay.asset.publisher.web.internal.helper.AssetRSSHelper;
 import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizerRegistry;
 import com.liferay.asset.util.AssetHelper;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
@@ -32,12 +35,17 @@ import com.liferay.dynamic.data.mapping.util.DDMUtil;
 import com.liferay.fragment.processor.PortletRegistry;
 import com.liferay.info.item.InfoItemServiceRegistry;
 import com.liferay.item.selector.ItemSelector;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.exception.NoSuchGroupException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.Release;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
@@ -50,10 +58,19 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.DateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Http;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.rss.export.RSSExporter;
+import com.liferay.rss.model.SyndContent;
+import com.liferay.rss.model.SyndEntry;
+import com.liferay.rss.model.SyndFeed;
+import com.liferay.rss.model.SyndLink;
+import com.liferay.rss.model.SyndModelFactory;
+import com.liferay.rss.util.RSSUtil;
 import com.liferay.segments.SegmentsEntryRetriever;
 import com.liferay.segments.context.RequestContextMapper;
 
@@ -63,14 +80,16 @@ import java.io.Serializable;
 
 import java.text.DateFormat;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
 import javax.portlet.PortletPreferences;
+import javax.portlet.PortletRequest;
+import javax.portlet.PortletResponse;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 import javax.portlet.ResourceRequest;
@@ -121,185 +140,6 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class AssetPublisherPortlet extends MVCPortlet {
 
-	public void getFieldValue(
-			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
-		throws PortletException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)resourceRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		try {
-			ServiceContext serviceContext = ServiceContextFactory.getInstance(
-				resourceRequest);
-
-			String className = ParamUtil.getString(
-				resourceRequest, "className");
-			long classTypeId = ParamUtil.getLong(
-				resourceRequest, "classTypeId");
-			String fieldName = ParamUtil.getString(resourceRequest, "name");
-
-			AssetRendererFactory<?> assetRendererFactory =
-				AssetRendererFactoryRegistryUtil.
-					getAssetRendererFactoryByClassName(className);
-
-			ClassTypeReader classTypeReader =
-				assetRendererFactory.getClassTypeReader();
-
-			ClassType classType = classTypeReader.getClassType(
-				classTypeId, themeDisplay.getLocale());
-
-			ClassTypeField classTypeField = classType.getClassTypeField(
-				fieldName);
-
-			Fields fields = (Fields)serviceContext.getAttribute(
-				Fields.class.getName() + classTypeField.getClassTypeId());
-
-			if (fields == null) {
-				String fieldsNamespace = ParamUtil.getString(
-					resourceRequest, "fieldsNamespace");
-
-				fields = DDMUtil.getFields(
-					classTypeField.getClassTypeId(), fieldsNamespace,
-					serviceContext);
-			}
-
-			Field field = fields.get(fieldName);
-
-			Serializable fieldValue = field.getValue(
-				themeDisplay.getLocale(), 0);
-
-			JSONObject jsonObject = jsonFactory.createJSONObject();
-
-			if (fieldValue != null) {
-				jsonObject.put("success", true);
-			}
-			else {
-				jsonObject.put("success", false);
-
-				writeJSON(resourceRequest, resourceResponse, jsonObject);
-
-				return;
-			}
-
-			jsonObject.put(
-				"displayValue", _getDisplayFieldValue(field, themeDisplay)
-			).put(
-				"value",
-				() -> {
-					if (fieldValue instanceof Boolean) {
-						return (Boolean)fieldValue;
-					}
-
-					if (fieldValue instanceof Date) {
-						DateFormat dateFormat =
-							DateFormatFactoryUtil.getSimpleDateFormat(
-								"yyyyMM ddHHmmss");
-
-						return dateFormat.format(fieldValue);
-					}
-
-					if (fieldValue instanceof Double) {
-						return (Double)fieldValue;
-					}
-
-					if (fieldValue instanceof Float) {
-						return (Float)fieldValue;
-					}
-
-					if (fieldValue instanceof Integer) {
-						return (Integer)fieldValue;
-					}
-
-					if (fieldValue instanceof Number) {
-						return String.valueOf(fieldValue);
-					}
-
-					return (String)fieldValue;
-				}
-			);
-
-			writeJSON(resourceRequest, resourceResponse, jsonObject);
-		}
-		catch (Exception exception) {
-			throw new PortletException(exception);
-		}
-	}
-
-	public void getRSS(
-			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
-		throws IOException {
-
-		PortletPreferences portletPreferences =
-			resourceRequest.getPreferences();
-
-		boolean enableRss = GetterUtil.getBoolean(
-			portletPreferences.getValue("enableRss", null));
-
-		if (!portal.isRSSFeedsEnabled() || !enableRss) {
-			try {
-				portal.sendRSSFeedsDisabledError(
-					resourceRequest, resourceResponse);
-			}
-			catch (ServletException servletException) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(servletException);
-				}
-			}
-
-			return;
-		}
-
-		String currentURL = portal.getCurrentURL(resourceRequest);
-
-		String cacheability = HttpComponentsUtil.getParameter(
-			currentURL, "p_p_cacheability", false);
-
-		if (cacheability.equals(ResourceURL.FULL)) {
-			HttpServletResponse httpServletResponse =
-				portal.getHttpServletResponse(resourceResponse);
-
-			String redirectURL = HttpComponentsUtil.removeParameter(
-				currentURL, "p_p_cacheability");
-
-			httpServletResponse.sendRedirect(redirectURL);
-
-			return;
-		}
-
-		resourceResponse.setContentType(ContentTypes.TEXT_XML_UTF8);
-
-		try (OutputStream outputStream =
-				resourceResponse.getPortletOutputStream()) {
-
-			String rootPortletId = PortletIdCodec.decodePortletName(
-				portal.getPortletId(resourceRequest));
-
-			AssetPublisherDisplayContext assetPublisherDisplayContext =
-				new AssetPublisherDisplayContext(
-					assetHelper, assetListAssetEntryProvider,
-					assetListEntrySegmentsEntryRelLocalService,
-					assetPublisherCustomizerRegistry.
-						getAssetPublisherCustomizer(rootPortletId),
-					assetPublisherHelper, assetPublisherWebConfiguration,
-					assetPublisherWebHelper, infoItemServiceRegistry,
-					itemSelector, portal, resourceRequest, resourceResponse,
-					resourceRequest.getPreferences(), requestContextMapper,
-					segmentsEntryRetriever);
-
-			resourceRequest.setAttribute(
-				AssetPublisherWebKeys.ASSET_PUBLISHER_DISPLAY_CONTEXT,
-				assetPublisherDisplayContext);
-
-			byte[] bytes = assetRSSHelper.getRSS(
-				resourceRequest, resourceResponse);
-
-			outputStream.write(bytes);
-		}
-		catch (Exception exception) {
-			_log.error("Unable to get RSS feed", exception);
-		}
-	}
-
 	@Override
 	public void render(
 			RenderRequest renderRequest, RenderResponse renderResponse)
@@ -321,38 +161,14 @@ public class AssetPublisherPortlet extends MVCPortlet {
 			resourceRequest.getResourceID());
 
 		if (resourceID.equals("getFieldValue")) {
-			getFieldValue(resourceRequest, resourceResponse);
+			_getFieldValue(resourceRequest, resourceResponse);
 		}
 		else if (resourceID.equals("getRSS")) {
-			getRSS(resourceRequest, resourceResponse);
+			_getRSS(resourceRequest, resourceResponse);
 		}
 		else {
 			super.serveResource(resourceRequest, resourceResponse);
 		}
-	}
-
-	public void subscribe(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		assetPublisherWebHelper.subscribe(
-			themeDisplay.getPermissionChecker(), themeDisplay.getScopeGroupId(),
-			themeDisplay.getPlid(), themeDisplay.getPpid());
-	}
-
-	public void unsubscribe(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		assetPublisherWebHelper.unsubscribe(
-			themeDisplay.getPermissionChecker(), themeDisplay.getPlid(),
-			themeDisplay.getPpid());
 	}
 
 	@Activate
@@ -465,9 +281,6 @@ public class AssetPublisherPortlet extends MVCPortlet {
 	protected AssetPublisherWebHelper assetPublisherWebHelper;
 
 	@Reference
-	protected AssetRSSHelper assetRSSHelper;
-
-	@Reference
 	protected InfoItemServiceRegistry infoItemServiceRegistry;
 
 	@Reference
@@ -492,6 +305,161 @@ public class AssetPublisherPortlet extends MVCPortlet {
 
 	@Reference
 	protected SegmentsEntryRetriever segmentsEntryRetriever;
+
+	private String _exportToRSS(
+			PortletRequest portletRequest, PortletResponse portletResponse,
+			String name, String format, double version, String displayStyle,
+			String linkBehavior, List<AssetEntry> assetEntries)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		SyndFeed syndFeed = _syndModelFactory.createSyndFeed();
+
+		syndFeed.setDescription(name);
+
+		List<SyndEntry> syndEntries = new ArrayList<>();
+
+		syndFeed.setEntries(syndEntries);
+
+		for (AssetEntry assetEntry : assetEntries) {
+			SyndEntry syndEntry = _syndModelFactory.createSyndEntry();
+
+			syndEntry.setAuthor(_portal.getUserName(assetEntry));
+
+			SyndContent syndContent = _syndModelFactory.createSyndContent();
+
+			syndContent.setType(RSSUtil.ENTRY_TYPE_DEFAULT);
+
+			String value = null;
+
+			String languageId = _language.getLanguageId(portletRequest);
+
+			if (displayStyle.equals(RSSUtil.DISPLAY_STYLE_TITLE)) {
+				value = StringPool.BLANK;
+			}
+			else {
+				value = assetEntry.getSummary(languageId, true);
+			}
+
+			syndContent.setValue(value);
+
+			syndEntry.setDescription(syndContent);
+
+			String link = _getEntryURL(
+				portletRequest, portletResponse, linkBehavior, assetEntry);
+
+			syndEntry.setLink(link);
+
+			syndEntry.setPublishedDate(assetEntry.getPublishDate());
+			syndEntry.setTitle(assetEntry.getTitle(languageId, true));
+			syndEntry.setUpdatedDate(assetEntry.getModifiedDate());
+			syndEntry.setUri(link);
+
+			syndEntries.add(syndEntry);
+		}
+
+		syndFeed.setFeedType(RSSUtil.getFeedType(format, version));
+
+		List<SyndLink> syndLinks = new ArrayList<>();
+
+		syndFeed.setLinks(syndLinks);
+
+		SyndLink selfSyndLink = _syndModelFactory.createSyndLink();
+
+		syndLinks.add(selfSyndLink);
+
+		String feedURL = _getFeedURL(portletRequest);
+
+		selfSyndLink.setHref(feedURL);
+
+		selfSyndLink.setRel("self");
+
+		SyndLink alternateSyndLink = _syndModelFactory.createSyndLink();
+
+		syndLinks.add(alternateSyndLink);
+
+		alternateSyndLink.setHref(_portal.getLayoutFullURL(themeDisplay));
+		alternateSyndLink.setRel("alternate");
+
+		syndFeed.setPublishedDate(new Date());
+		syndFeed.setTitle(name);
+		syndFeed.setUri(feedURL);
+
+		return _rssExporter.export(syndFeed);
+	}
+
+	private List<AssetEntry> _getAssetEntries(
+			PortletRequest portletRequest,
+			PortletPreferences portletPreferences)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		List<AssetEntry> assetEntries = new ArrayList<>();
+
+		SearchContainer<AssetEntry> searchContainer = new SearchContainer();
+
+		AssetPublisherDisplayContext assetPublisherDisplayContext =
+			(AssetPublisherDisplayContext)portletRequest.getAttribute(
+				AssetPublisherWebKeys.ASSET_PUBLISHER_DISPLAY_CONTEXT);
+
+		searchContainer.setDelta(assetPublisherDisplayContext.getRSSDelta());
+
+		Map<String, Serializable> attributes =
+			assetPublisherDisplayContext.getAttributes();
+
+		attributes.put("filterExpired", Boolean.TRUE);
+
+		List<AssetEntryResult> assetEntryResults =
+			_assetPublisherHelper.getAssetEntryResults(
+				searchContainer,
+				assetPublisherDisplayContext.getAssetEntryQuery(),
+				themeDisplay.getLayout(), portletPreferences,
+				assetPublisherDisplayContext.getPortletName(),
+				themeDisplay.getLocale(), themeDisplay.getTimeZone(),
+				themeDisplay.getCompanyId(), themeDisplay.getScopeGroupId(),
+				themeDisplay.getUserId(),
+				assetPublisherDisplayContext.getClassNameIds(), attributes);
+
+		for (AssetEntryResult assetEntryResult : assetEntryResults) {
+			assetEntries.addAll(assetEntryResult.getAssetEntries());
+		}
+
+		return assetEntries;
+	}
+
+	private String _getAssetPublisherURL(PortletRequest portletRequest)
+		throws Exception {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		Layout layout = themeDisplay.getLayout();
+
+		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
+
+		StringBundler sb = new StringBundler(6);
+
+		String layoutFriendlyURL = GetterUtil.getString(
+			_portal.getLayoutFriendlyURL(layout, themeDisplay));
+
+		if (!layoutFriendlyURL.startsWith(Http.HTTP_WITH_SLASH) &&
+			!layoutFriendlyURL.startsWith(Http.HTTPS_WITH_SLASH)) {
+
+			sb.append(themeDisplay.getPortalURL());
+		}
+
+		sb.append(layoutFriendlyURL);
+		sb.append(Portal.FRIENDLY_URL_SEPARATOR);
+		sb.append("asset_publisher/");
+		sb.append(portletDisplay.getInstanceId());
+		sb.append(StringPool.SLASH);
+
+		return sb.toString();
+	}
 
 	private String _getDisplayFieldValue(Field field, ThemeDisplay themeDisplay)
 		throws Exception {
@@ -519,9 +487,293 @@ public class AssetPublisherPortlet extends MVCPortlet {
 		return fieldValue;
 	}
 
+	private String _getEntryURL(
+			PortletRequest portletRequest, PortletResponse portletResponse,
+			String linkBehavior, AssetEntry assetEntry)
+		throws Exception {
+
+		if (linkBehavior.equals("viewInPortlet")) {
+			return _getEntryURLViewInContext(
+				portletRequest, portletResponse, assetEntry);
+		}
+
+		return _getEntryURLAssetPublisher(portletRequest, assetEntry);
+	}
+
+	private String _getEntryURLAssetPublisher(
+			PortletRequest portletRequest, AssetEntry assetEntry)
+		throws Exception {
+
+		AssetRendererFactory<?> assetRendererFactory =
+			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
+				assetEntry.getClassName());
+
+		return StringBundler.concat(
+			_getAssetPublisherURL(portletRequest),
+			assetRendererFactory.getType(), "/id/", assetEntry.getEntryId());
+	}
+
+	private String _getEntryURLViewInContext(
+		PortletRequest portletRequest, PortletResponse portletResponse,
+		AssetEntry assetEntry) {
+
+		String assetViewURL = _assetPublisherHelper.getAssetViewURL(
+			_portal.getLiferayPortletRequest(portletRequest),
+			_portal.getLiferayPortletResponse(portletResponse), assetEntry,
+			true);
+
+		if (Validator.isNotNull(assetViewURL) &&
+			!assetViewURL.startsWith(Http.HTTP_WITH_SLASH) &&
+			!assetViewURL.startsWith(Http.HTTPS_WITH_SLASH)) {
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)portletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			assetViewURL = themeDisplay.getPortalURL() + assetViewURL;
+		}
+
+		return assetViewURL;
+	}
+
+	private String _getFeedURL(PortletRequest portletRequest) throws Exception {
+		String feedURL = _getAssetPublisherURL(portletRequest);
+
+		return feedURL.concat("rss");
+	}
+
+	private void _getFieldValue(
+			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
+		throws PortletException {
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)resourceRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+				resourceRequest);
+
+			String className = ParamUtil.getString(
+				resourceRequest, "className");
+			long classTypeId = ParamUtil.getLong(
+				resourceRequest, "classTypeId");
+			String fieldName = ParamUtil.getString(resourceRequest, "name");
+
+			AssetRendererFactory<?> assetRendererFactory =
+				AssetRendererFactoryRegistryUtil.
+					getAssetRendererFactoryByClassName(className);
+
+			ClassTypeReader classTypeReader =
+				assetRendererFactory.getClassTypeReader();
+
+			ClassType classType = classTypeReader.getClassType(
+				classTypeId, themeDisplay.getLocale());
+
+			ClassTypeField classTypeField = classType.getClassTypeField(
+				fieldName);
+
+			Fields fields = (Fields)serviceContext.getAttribute(
+				Fields.class.getName() + classTypeField.getClassTypeId());
+
+			if (fields == null) {
+				String fieldsNamespace = ParamUtil.getString(
+					resourceRequest, "fieldsNamespace");
+
+				fields = DDMUtil.getFields(
+					classTypeField.getClassTypeId(), fieldsNamespace,
+					serviceContext);
+			}
+
+			Field field = fields.get(fieldName);
+
+			Serializable fieldValue = field.getValue(
+				themeDisplay.getLocale(), 0);
+
+			JSONObject jsonObject = jsonFactory.createJSONObject();
+
+			if (fieldValue != null) {
+				jsonObject.put("success", true);
+			}
+			else {
+				jsonObject.put("success", false);
+
+				writeJSON(resourceRequest, resourceResponse, jsonObject);
+
+				return;
+			}
+
+			jsonObject.put(
+				"displayValue", _getDisplayFieldValue(field, themeDisplay)
+			).put(
+				"value",
+				() -> {
+					if (fieldValue instanceof Boolean) {
+						return (Boolean)fieldValue;
+					}
+
+					if (fieldValue instanceof Date) {
+						DateFormat dateFormat =
+							DateFormatFactoryUtil.getSimpleDateFormat(
+								"yyyyMM ddHHmmss");
+
+						return dateFormat.format(fieldValue);
+					}
+
+					if (fieldValue instanceof Double) {
+						return (Double)fieldValue;
+					}
+
+					if (fieldValue instanceof Float) {
+						return (Float)fieldValue;
+					}
+
+					if (fieldValue instanceof Integer) {
+						return (Integer)fieldValue;
+					}
+
+					if (fieldValue instanceof Number) {
+						return String.valueOf(fieldValue);
+					}
+
+					return (String)fieldValue;
+				}
+			);
+
+			writeJSON(resourceRequest, resourceResponse, jsonObject);
+		}
+		catch (Exception exception) {
+			throw new PortletException(exception);
+		}
+	}
+
+	private void _getRSS(
+			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
+		throws IOException {
+
+		PortletPreferences portletPreferences =
+			resourceRequest.getPreferences();
+
+		boolean enableRss = GetterUtil.getBoolean(
+			portletPreferences.getValue("enableRss", null));
+
+		if (!portal.isRSSFeedsEnabled() || !enableRss) {
+			try {
+				portal.sendRSSFeedsDisabledError(
+					resourceRequest, resourceResponse);
+			}
+			catch (ServletException servletException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(servletException);
+				}
+			}
+
+			return;
+		}
+
+		String currentURL = portal.getCurrentURL(resourceRequest);
+
+		String cacheability = HttpComponentsUtil.getParameter(
+			currentURL, "p_p_cacheability", false);
+
+		if (cacheability.equals(ResourceURL.FULL)) {
+			HttpServletResponse httpServletResponse =
+				portal.getHttpServletResponse(resourceResponse);
+
+			String redirectURL = HttpComponentsUtil.removeParameter(
+				currentURL, "p_p_cacheability");
+
+			httpServletResponse.sendRedirect(redirectURL);
+
+			return;
+		}
+
+		resourceResponse.setContentType(ContentTypes.TEXT_XML_UTF8);
+
+		try (OutputStream outputStream =
+				resourceResponse.getPortletOutputStream()) {
+
+			String rootPortletId = PortletIdCodec.decodePortletName(
+				portal.getPortletId(resourceRequest));
+
+			AssetPublisherDisplayContext assetPublisherDisplayContext =
+				new AssetPublisherDisplayContext(
+					assetHelper, assetListAssetEntryProvider,
+					assetListEntrySegmentsEntryRelLocalService,
+					assetPublisherCustomizerRegistry.
+						getAssetPublisherCustomizer(rootPortletId),
+					assetPublisherHelper, assetPublisherWebConfiguration,
+					assetPublisherWebHelper, infoItemServiceRegistry,
+					itemSelector, portal, resourceRequest, resourceResponse,
+					resourceRequest.getPreferences(), requestContextMapper,
+					segmentsEntryRetriever);
+
+			resourceRequest.setAttribute(
+				AssetPublisherWebKeys.ASSET_PUBLISHER_DISPLAY_CONTEXT,
+				assetPublisherDisplayContext);
+
+			byte[] bytes = _getRSSBytes(resourceRequest, resourceResponse);
+
+			outputStream.write(bytes);
+		}
+		catch (Exception exception) {
+			_log.error("Unable to get RSS feed", exception);
+		}
+	}
+
+	private byte[] _getRSSBytes(
+			ResourceRequest portletRequest, ResourceResponse portletResponse)
+		throws Exception {
+
+		PortletPreferences portletPreferences = portletRequest.getPreferences();
+
+		String selectionStyle = portletPreferences.getValue(
+			"selectionStyle",
+			AssetPublisherSelectionStyleConfigurationUtil.
+				defaultSelectionStyle());
+
+		if (!selectionStyle.equals(
+				AssetPublisherSelectionStyleConstants.TYPE_DYNAMIC)) {
+
+			return new byte[0];
+		}
+
+		String assetLinkBehavior = portletPreferences.getValue(
+			"assetLinkBehavior", "showFullContent");
+		String rssDisplayStyle = portletPreferences.getValue(
+			"rssDisplayStyle", RSSUtil.DISPLAY_STYLE_ABSTRACT);
+		String rssFeedType = portletPreferences.getValue(
+			"rssFeedType", RSSUtil.FEED_TYPE_DEFAULT);
+		String rssName = portletPreferences.getValue("rssName", null);
+
+		String format = RSSUtil.getFeedTypeFormat(rssFeedType);
+		double version = RSSUtil.getFeedTypeVersion(rssFeedType);
+
+		String rss = _exportToRSS(
+			portletRequest, portletResponse, rssName, format, version,
+			rssDisplayStyle, assetLinkBehavior,
+			_getAssetEntries(portletRequest, portletPreferences));
+
+		return rss.getBytes(StringPool.UTF8);
+	}
+
 	private static final String _ALIAS = "asset-list";
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		AssetPublisherPortlet.class);
+
+	@Reference
+	private AssetPublisherHelper _assetPublisherHelper;
+
+	@Reference
+	private Language _language;
+
+	@Reference
+	private Portal _portal;
+
+	@Reference
+	private RSSExporter _rssExporter;
+
+	@Reference
+	private SyndModelFactory _syndModelFactory;
 
 }
